[新需求]: 变量初始化
Issue: https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/IC9PPD

+++ ecmascript/interpreter/interpreter_assembly.cpp
@@ -298,7 +298,7 @@ void InterpreterAssembly::MethodEntry(JSThread *thread, Method *method, JSTagged
 
 int64_t InterpreterAssembly::GetCallSize(EcmaOpcode opcode)
 {
-    int64_t callSize = BytecodeInstruction::Size(opcode);
+    int64_t callSize = static_cast<int64_t>(BytecodeInstruction::Size(opcode));
     switch (opcode) {
         case EcmaOpcode::SUPERCALLSPREAD_IMM8_V8:
         case EcmaOpcode::SUPERCALLTHISRANGE_IMM8_IMM8_V8:
